Refactor delete vector storage to process in batches

+++ src/domain/services/DataSourceManagementService.ts
@@ -180,6 +180,7 @@ export class DataSourceManagementService {
       documentsBlobUrl,
       metadata,
       indexPercentage,
+      isBlobStorageDeleted,
       ...rest
     } = knowledge;
 
@@ -195,6 +196,7 @@ export class DataSourceManagementService {
       documentsBlobUrl,
       indexPercentage: indexPercentage.toString(),
       metadata,
+      isBlobStorageDeleted,
     };
   }
 
@@ -1132,6 +1134,10 @@ export class DataSourceManagementService {
 
   public async deleteBlobStorage(knowledgeId: string) {
     const knowledge = await this.knowledgeRepository.getById(knowledgeId);
+    if (knowledge.isBlobStorageDeleted) {
+      return;
+    }
+
     try {
       if (knowledge.documentsBlobUrl) {
         await FileStorageService.delete(knowledge.documentsBlobUrl);
